Move externally provided nvComp
- makes nvComp a dlopen depenency of DALI
- as nvComp provides a wheel this PR makes DALI dependent on
  it so we don't need to bundle it with DALI anymore
