+++ conda/dali_native_libs/recipe/build.sh
@@ -93,6 +93,7 @@ cmake -DCUDA_TOOLKIT_ROOT_DIR=/usr/local/cuda \
       -DWITH_DYNAMIC_CUFFT=${WITH_DYNAMIC_CUFFT:-ON}       \
       -DWITH_DYNAMIC_NPP=${WITH_DYNAMIC_NPP:-ON}           \
       -DWITH_DYNAMIC_NVIMGCODEC=${WITH_DYNAMIC_NVIMGCODEC:-ON} \
+      -DWITH_DYNAMIC_NVCOMP=${DWITH_DYNAMIC_NVCOMP:-ON  } \
       -DVERBOSE_LOGS=${VERBOSE_LOGS:-OFF}                 \
       -DWERROR=${WERROR:-ON}                              \
       -DBUILD_WITH_ASAN=${BUILD_WITH_ASAN:-OFF}           \
@@ -125,11 +126,9 @@ DEPS_LIST=(
 
 DEPS_PATH=/usr/local/
 
-if [ "$BUILD_NVCOMP" = "ON" ]; then
+if [ "$BUILD_NVCOMP" = "ON" ] && [ "$WITH_DYNAMIC_NVCOMP" = "ON" ]; then
     DEPS_LIST+=(
-        "${DEPS_PATH}/cuda/lib64/libnvcomp.so"
-        "${DEPS_PATH}/cuda/lib64/libnvcomp_gdeflate.so"
-        "${DEPS_PATH}/cuda/lib64/libnvcomp_bitcomp.so"
+        "${DEPS_PATH}/cuda/lib64/libnvcomp.so.4"
     )
 fi
