+++ dali/python/setup.py.in
@@ -91,6 +91,7 @@ For more details please check the
           'dm-tree <= 0.1.9; python_version>="3.10"',
           'packaging <= 24.2',
           @DALI_INSTALL_REQUIRES_NVIMGCODEC@
+          @DALI_INSTALL_REQUIRES_NVCOMP@
           ],
      )
