+++ docker/Dockerfile
@@ -94,6 +94,8 @@ ARG WITH_DYNAMIC_NPP
 ENV WITH_DYNAMIC_NPP=${WITH_DYNAMIC_NPP:-ON}
 ARG WITH_DYNAMIC_NVIMGCODEC
 ENV WITH_DYNAMIC_NVIMGCODEC=${WITH_DYNAMIC_NVIMGCODEC:-ON}
+ARG WITH_DYNAMIC_NVCOMP
+ENV WITH_DYNAMIC_NVCOMP=${WITH_DYNAMIC_NVCOMP:-ON}
 ARG BUILD_WITH_ASAN
 ENV BUILD_WITH_ASAN=${BUILD_WITH_ASAN}
 ENV BUILD_WITH_LSAN=${BUILD_WITH_LSAN}
